refactor: 도메인 로직 모듈 분리
