--- app/src/main/java/woowacourse/movie/domain/discount/DiscountRule.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.domain.discount
-
-import woowacourse.movie.domain.payment.PaymentAmount
-import java.time.LocalDateTime
-
-interface DiscountRule {
-    fun getPaymentAmountResult(paymentAmount: PaymentAmount, screeningDateTime: LocalDateTime): PaymentAmount
-}
