--- app/src/main/java/woowacourse/movie/domain/movie/Movie.kt
@@ -1,8 +1,0 @@
-package woowacourse.movie.domain.movie
-
-data class Movie(
-    val name: Name,
-    val screeningPeriod: ScreeningPeriod,
-    val runningTime: Int,
-    val description: String
-)
