--- app/src/main/java/woowacourse/movie/domain/movie/Name.kt
@@ -1,14 +1,0 @@
-package woowacourse.movie.domain.movie
-
-@JvmInline
-value class Name(
-    val value: String
-) {
-    init {
-        require(value.length >= MINIMUM_LENGTH)
-    }
-
-    companion object {
-        private const val MINIMUM_LENGTH = 1
-    }
-}
