--- app/src/main/java/woowacourse/movie/domain/payment/PaymentAmount.kt
@@ -1,12 +1,0 @@
-package woowacourse.movie.domain.payment
-
-@JvmInline
-value class PaymentAmount(val value: Int) {
-    init {
-        require(value >= MINIMUM)
-    }
-
-    companion object {
-        private const val MINIMUM = 0
-    }
-}
