--- app/src/main/java/woowacourse/movie/domain/payment/PaymentType.kt
@@ -1,5 +1,0 @@
-package woowacourse.movie.domain.payment
-
-enum class PaymentType {
-    LOCAL_PAYMENT;
-}
