--- app/src/main/java/woowacourse/movie/domain/reservation/TicketCount.kt
@@ -1,13 +1,0 @@
-package woowacourse.movie.domain.reservation
-
-@JvmInline
-value class TicketCount(val value: Int) {
-
-    init {
-        require(value >= MINIMUM)
-    }
-
-    companion object {
-        const val MINIMUM = 1
-    }
-}
