+++ domain/build.gradle.kts
@@ -1,0 +1,26 @@
+plugins {
+    kotlin("jvm")
+//    id("java-library")
+//    id("org.jetbrains.kotlin.jvm")
+}
+
+dependencies {
+    testImplementation("org.junit.jupiter", "junit-jupiter", "5.8.2")
+    testImplementation("org.assertj", "assertj-core", "3.22.0")
+    testImplementation("io.kotest", "kotest-runner-junit5", "5.2.3")
+}
+
+tasks {
+    compileKotlin {
+        kotlinOptions.jvmTarget = "11"
+    }
+    compileTestKotlin {
+        kotlinOptions.jvmTarget = "11"
+    }
+    test {
+        useJUnitPlatform()
+    }
+    ktlint {
+        verbose.set(true)
+    }
+}
