+++ domain/src/main/kotlin/discount/Discount.kt
@@ -1,0 +1,23 @@
+package discount
+
+import payment.PaymentAmount
+import java.time.LocalDateTime
+
+class Discount(
+    private val discountRules: List<DiscountRule>
+) : DiscountRule {
+
+    constructor(vararg rules: DiscountRule) : this(rules.map { it }.toList())
+
+    override fun getPaymentAmountResult(
+        paymentAmount: PaymentAmount,
+        screeningDateTime: LocalDateTime
+    ): PaymentAmount {
+
+        var resultPaymentAmount: PaymentAmount = paymentAmount
+        discountRules.forEach {
+            resultPaymentAmount = it.getPaymentAmountResult(resultPaymentAmount, screeningDateTime)
+        }
+        return resultPaymentAmount
+    }
+}
