+++ domain/src/main/kotlin/discount/DiscountRule.kt
@@ -1,0 +1,8 @@
+package discount
+
+import payment.PaymentAmount
+import java.time.LocalDateTime
+
+interface DiscountRule {
+    fun getPaymentAmountResult(paymentAmount: PaymentAmount, screeningDateTime: LocalDateTime): PaymentAmount
+}
