+++ domain/src/main/kotlin/movie/Movie.kt
@@ -1,0 +1,8 @@
+package movie
+
+data class Movie(
+    val name: Name,
+    val screeningPeriod: ScreeningPeriod,
+    val runningTime: Int,
+    val description: String
+)
