+++ domain/src/main/kotlin/payment/PaymentAmount.kt
@@ -1,0 +1,12 @@
+package payment
+
+@JvmInline
+value class PaymentAmount(val value: Int) {
+    init {
+        require(value >= MINIMUM)
+    }
+
+    companion object {
+        private const val MINIMUM = 0
+    }
+}
