+++ domain/src/main/kotlin/payment/PaymentType.kt
@@ -1,0 +1,5 @@
+package payment
+
+enum class PaymentType {
+    LOCAL_PAYMENT;
+}
