+++ domain/src/main/kotlin/reservation/TicketCount.kt
@@ -1,0 +1,13 @@
+package reservation
+
+@JvmInline
+value class TicketCount(val value: Int) {
+
+    init {
+        require(value >= MINIMUM)
+    }
+
+    companion object {
+        const val MINIMUM = 1
+    }
+}
